hits-bot/requirements.txt
aiogram==3.4.1
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
aiosqlite==0.19.0
jinja2==3.1.3
aiohttp==3.9.5
python-multipart==0.0.9
